refactoring out TimeControls
